Extract KalmanFilter state and covariance into types
