+++ crates/filtering/src/lib.rs
@@ -9,6 +9,6 @@ mod tap_detector;
 pub use hysteresis::{greater_than_with_hysteresis, less_than_with_hysteresis};
 pub use kalman_filter::KalmanFilter;
 pub use low_pass_filter::LowPassFilter;
-pub use orientation_filter::{OrientationFilter, OrientationFilterParameters};
+pub use orientation_filter::OrientationFilter;
 pub use pose_filter::{PoseFilter, ScoredPoseFilter};
 pub use tap_detector::TapDetector;
