+++ crates/types/Cargo.toml
@@ -7,7 +7,6 @@ license = "GPL-3.0-only"
 [dependencies]
 approx = { workspace = true }
 color-eyre = { workspace = true }
-filtering = { workspace = true }
 image = { workspace = true }
 nalgebra = { workspace = true }
 ordered-float = { workspace = true }
