+++ crates/types/src/ball_filter_hypothesis.rs
@@ -1,11 +1,12 @@
 use std::time::SystemTime;
 
-use filtering::KalmanFilter;
 use serde::{Deserialize, Serialize};
+
+use crate::kalman_filter::KalmanFilterSnapshot;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BallFilterHypothesis {
-    pub filter: KalmanFilter<4>,
+    pub filter: KalmanFilterSnapshot<4>,
     pub validity: f32,
     pub last_update: SystemTime,
 }
