+++ crates/types/src/lib.rs
@@ -23,6 +23,7 @@ pub mod image;
 mod image_segments;
 mod initial_pose;
 mod joints;
+pub mod kalman_filter;
 mod kick_decision;
 mod kick_step;
 mod led;
@@ -38,6 +39,7 @@ mod motion_file;
 mod motion_selection;
 mod obstacle_filter_hypothesis;
 mod obstacles;
+pub mod orientation_filter;
 mod path_obstacles;
 mod penalty_shot_direction;
 mod perspective_grid_candidates;
