+++ crates/types/src/obstacle_filter_hypothesis.rs
@@ -1,13 +1,12 @@
 use std::time::SystemTime;
 
-use filtering::KalmanFilter;
 use serde::{Deserialize, Serialize};
 
-use crate::ObstacleKind;
+use crate::{kalman_filter::KalmanFilterSnapshot, ObstacleKind};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ObstacleFilterHypothesis {
-    pub filter: KalmanFilter<2>,
+    pub filter: KalmanFilterSnapshot<2>,
     pub measurement_count: usize,
     pub last_update: SystemTime,
     pub obstacle_kind: ObstacleKind,
